requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.5
pillow>=10.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0
torch>=2.0.0
torchvision>=0.15.0
google-generativeai>=0.8.0
reportlab>=4.0.0
